arxiv-agent/requirements.txt
requests
gradio
groq
python-dotenv